requirements.txt
langchain>=0.1.0
langchain-openai
langgraph
redis
chromadb
pydantic
python-dotenv
tiktoken
